Update version 1.0

+++ app/src/main/java/com/rztechtunes/dailyexpensemanager/dao/ExpenseIncomeDao.java
@@ -22,7 +22,7 @@ public interface ExpenseIncomeDao {
     int deleteExIncome(ExpenseIncomePojo expenseIncomePojo);
 
     @Query("Select * from expenseincometbl where e_id like:id")
-    ExpenseIncomePojo getAllDatabyId(int id);
+    ExpenseIncomePojo getAllDatabyId(long id);
 
     @Query("Select * from expenseincometbl order by e_id desc")
     List<ExpenseIncomePojo> getAllData();
